// mudanzas-romeroj-app/src/app/types/Presupuesto.ts
export class Presupuesto {
  
  tipoServicio: string = '';
  nombre: string = '';
  telefono: string = '';
  email: string = '';
  fecha: string = ''; 
  direccionOrigen: string = '';
  direccionDestino: string = '';
  consulta?: string = '';
  politicaPrivacidad: boolean = false;

  constructor(){}
}